\name{topVar.mcia}
\alias{topVar.mcia}

\title{
Selecting top weighted variables (genes) from result of MCIA
}
\description{
  This function provides a method selecting 
  top weighted variables (genes) on an axis (either positive side or negative side or both)
  from an object of class \code{mcia}.
}
\usage{
\method{topVar}{mcia}(x, axis = 1, end = "both", topN = 5)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    See \code{\link{plotVar.mcia}}
}
  \item{axis}{
    See \code{\link{plotVar.mcia}}
}
  \item{end}{
    See \code{\link{plotVar.mcia}}
}
  \item{topN}{
    See \code{\link{plotVar.mcia}}
}
}

\value{
    See \code{\link{plotVar.mcia}}
}

\author{
    Chen Meng
}


